[Android/Feat] 루틴 등록 기능 구현 - 공기청정기 S12P31D103-210

See merge request s12-final/S12P31D103!142

+++ Android/data/src/main/java/com/hogumiwarts/data/repository/AirpurifierRepositoryImpl.kt
@@ -12,10 +12,10 @@ import com.hogumiwarts.domain.model.airpurifier.PatchAirpurifierPowerResult
 import com.hogumiwarts.domain.repository.AirpurifierRepository
 import javax.inject.Inject
 
-class AirpurifierRepositoryImpl@Inject constructor(
+class AirpurifierRepositoryImpl @Inject constructor(
     private val airpurifierApi: AirpurifierApi // 🔹 Retrofit API 인터페이스 주입
 ):AirpurifierRepository {
-    override suspend fun getAirpurifierStatus(deviceId: Long): AirpurifierResult {
+    override suspend fun getAirpurifierStatus(deviceId: Int): AirpurifierResult {
         return try {
             // ✅ API 호출
             val response = airpurifierApi.getAirpurifierStatus(deviceId)
@@ -40,7 +40,7 @@ class AirpurifierRepositoryImpl@Inject constructor(
     }
 
     override suspend fun patchAirpurifierPower(
-        deviceId: Long,
+        deviceId: Int,
         activated: Boolean,
     ): PatchAirpurifierPowerResult {
         return try {
@@ -67,7 +67,7 @@ class AirpurifierRepositoryImpl@Inject constructor(
     }
 
     override suspend fun patchAirpurifierFanMode(
-        deviceId: Long,
+        deviceId: Int,
         fanMode: String,
     ): PatchAirpurifierFanModeResult {
         return try {
